+++ grails-app/init/io/xh/hoist/BootStrap.groovy
@@ -128,6 +128,13 @@ class BootStrap {
                 groupName: 'xh.io',
                 note: 'True to enable the log viewer included with the Hoist Admin console as well as the associated server-side endpoints.'
             ],
+            xhEnableMonitoring: [
+                    valueType: 'bool',
+                    defaultValue: true,
+                    clientVisible: true,
+                    groupName: 'xh.io',
+                    note: 'True to enable the monitor tab included with the Hoist Admin console and the associated server-side jobs'
+            ],
             xhExportConfig: [
                 valueType: 'json',
                 defaultValue: [